requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
influxdb-client==1.36.1
msgpack==1.0.5
paho-mqtt==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reactivex==4.0.4
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.3
